[#77] path validation has been removed for export adapter
