+++ dicom_util/src/main/java/com/google/cloud/healthcare/DicomWebClientJetty.java
@@ -47,8 +47,6 @@ public class DicomWebClientJetty implements IDicomWebClient {
       String stowPath) {
     this.credentials = credentials;
     this.stowPath = StringUtil.trim(stowPath);
-
-    DicomWebValidation.validatePath(this.stowPath, DicomWebValidation.DICOMWEB_STUDIES_VALIDATION);
   }
 
   @Override
